Derive paths from $MODDIR instead of hardcoding
Magisk v16.3 drops the backwards-compatibility symlinks.
In order to load sshd, we need to export the MODDIR variable to the init
script.
Also sshd sets the PATH for a new session. This path was before
hardcoded to use the compatibility symlink /magisk. Now, it also uses
the newly exported MODDIR to calculate the correct path without these
symlinks.